moment_kinetics is a package; modules are submodules of moment_kinetics
Make moment_kinetics a Julia package by moving source files to src/
directory and adding a Project.toml file. Also includes a Manifest.toml
file, which was created by Pkg when adding the dependencies.

Make all the defined modules submodules of the 'moment_kinetics' module.
Means modules can be accessed using relative paths, and we no longer
need `push!(LOAD_PATH, ".")` statements.
