+++ run_post_processing.jl
@@ -1,0 +1,5 @@
+import moment_kinetics as mk
+
+if abspath(PROGRAM_FILE) == @__FILE__
+    mk.post_processing.analyze_and_plot_data(ARGS[1])
+end
