+++ src/clenshaw_curtis.jl
