+++ src/vpa_advection.jl
@@ -3,12 +3,12 @@ module vpa_advection
 export vpa_advection!
 export update_speed_vpa!
 
-using semi_lagrange: find_approximate_characteristic!
-using advection: update_boundary_indices!
-using advection: advance_f_local!
-using em_fields: update_phi!
-using calculus: derivative!
-using initial_conditions: enforce_vpa_boundary_condition!
+using ..semi_lagrange: find_approximate_characteristic!
+using ..advection: update_boundary_indices!
+using ..advection: advance_f_local!
+using ..em_fields: update_phi!
+using ..calculus: derivative!
+using ..initial_conditions: enforce_vpa_boundary_condition!
 
 function vpa_advection!(f_out, fvec_in, ff, fields, moments, SL, advect,
 	vpa, z, use_semi_lagrange, dt, t, vpa_spectral, z_spectral, composition, CX_frequency, istage)
